requirements.txt
aiobotocore==2.3.4
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
ansi2html==1.8.0
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.3.0
astroid==2.12.5
astropy==4.0
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.3.0
attrs==21.4.0
autopep8==1.4.4
autovizwidget @ file:///home/conda/feedstock_root/build_artifacts/autovizwidget_1652515004592/work
awscli==1.25.63
Babel==2.10.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
bleach==5.0.1
blis==0.7.9
bokeh==1.4.0
boto==2.49.0
boto3==1.24.62
botocore==1.27.62
Bottleneck==1.3.2
Brotli==1.0.9
brotlipy==0.7.0
cached-property==1.5.2
catalogue==2.0.8
certifi==2022.6.15
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.1.0
clyent==1.2.2
colorama==0.4.3
conda==4.14.0
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649105789509/work
confection==0.0.3
contextlib2==0.6.0.post1
cryptography==37.0.4
cycler==0.10.0
cymem==2.0.7
Cython==0.29.15
cytoolz==0.10.1
dash==2.6.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2022.2.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.5.1
distlib==0.3.6
distributed==2022.2.0
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
dparse==0.5.2
ecdsa==0.13
emoji==2.2.0
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.8.0
flake8==3.7.9
Flask==1.1.1
Flask-Compress==1.12
frozenlist==1.3.1
fsspec==2022.7.1
funcy==1.17
future==0.18.2
gensim==4.2.0
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.50.0
h5py==2.10.0
hdijupyterutils @ file:///home/conda/feedstock_root/build_artifacts/hdijupyterutils_1652514846623/work
HeapDict==1.0.1
html5lib==1.0.1
http-ece==0.7.0
hypothesis==5.5.4
idna==2.7
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==4.12.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.34.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.1
jeepney==0.4.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyter-dash==0.4.2
jupyterlab==1.2.21
jupyterlab-server==1.0.6
keyring==21.1.0
kiwisolver==1.1.0
langcodes==3.3.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
llvmlite==0.39.0
locket==0.2.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.1.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
modin==0.12.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==1.0.4
multidict==6.0.2
multipledispatch==0.6.0
multiprocess==0.70.13
murmurhash==1.0.9
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.5.5
networkx==2.4
nltk==3.7
nose==1.3.7
notebook==6.4.12
numba==0.56.0
numexpr==2.7.1
numpy==1.21.6
numpydoc==0.9.2
olefile==0.46
openpyxl==3.0.3
packaging==20.1
pandas==1.3.5
pandocfilters==1.4.2
parso==0.8.3
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathos==0.2.9
pathtools==0.1.2
pathy==0.6.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.5.0.1
platformdirs==2.5.2
plotly==5.8.2
pluggy==0.13.1
ply==3.11
pox==0.3.1
ppft==1.7.6.5
preshed==3.0.8
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.20.1
protobuf3-to-dict==0.1.5
psutil==5.6.7
ptyprocess==0.6.0
pure-sasl @ file:///home/conda/feedstock_root/build_artifacts/pure-sasl_1631890804823/work
py==1.11.0
py-vapid==0.7.1
pyarrow==9.0.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
pydantic==1.10.2
pydocstyle==4.0.1
pyelliptic==1.5.7
pyflakes==2.1.1
pyfunctional==1.4.3
Pygments==2.5.2
PyHive @ file:///home/conda/feedstock_root/build_artifacts/pyhive_1646707521362/work
pykerberos==1.2.1
pyLDAvis==3.3.1
pylint==2.15.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-dotenv==0.20.0
python-jose==0.6.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==6.0
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
ray==2.1.0
regex==2022.8.17
requests==2.20.0
requests-kerberos @ file:///home/conda/feedstock_root/build_artifacts/requests-kerberos_1636815972022/work
retrying==1.3.3
rope==0.16.0
rsa==4.9
Rtree==0.9.3
ruamel_yaml==0.15.87
s3fs==2022.7.1
s3transfer==0.6.0
sagemaker==2.107.0
sagemaker-studio-analytics-extension==0.0.10
sagemaker-studio-sparkmagic-lib==0.1.4
sasl==0.2.1
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
Send2Trash==1.8.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
smart-open==5.2.1
smclarify==0.2
smdebug-rulesconfig==1.0.1
snowballstemmer==2.0.0
solidwebpush==1.2.3
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sparkmagic @ file:///home/conda/feedstock_root/build_artifacts/sparkmagic_1652120632036/work/sparkmagic
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
srsly==2.4.5
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tabulate==0.8.10
tblib==1.6.0
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1626090218611/work
terminado==0.8.3
testpath==0.4.4
texttable==1.6.4
thinc==8.1.5
thrift @ file:///home/conda/feedstock_root/build_artifacts/thrift_1614851254062/work
thrift-sasl @ file:///home/conda/feedstock_root/build_artifacts/thrift_sasl_1631902449705/work
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
toolz==0.10.0
tornado==6.2
tqdm==4.42.1
traitlets==4.3.3
typed-ast==1.5.4
typer==0.7.0
typing_extensions==4.1.1
ujson==5.4.0
unicodecsv==0.14.1
urllib3==1.24.3
virtualenv==20.16.7
wasabi==0.10.1
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
yapf==0.28.0
yarl==1.8.1
zict==1.0.0
zipp==3.8.1
